# ARROW-9075: [C++] Optimized Filter implementation: faster performance + compilation, smaller code size