… cpp/src/arrow/compute/kernels/vector_filter.cc
… cpp/src/arrow/compute/kernels/vector_filter_test.cc
… cpp/src/arrow/compute/kernels/vector_selection.cc
… cpp/src/arrow/compute/kernels/vector_selection_internal.h
… cpp/src/arrow/compute/kernels/vector_take.cc
… cpp/src/arrow/compute/kernels/vector_take_test.cc