@wesm 2020-06-16
Here's benchmark runs on my machine

* BEFORE: https://gist.github.com/wesm/857a3179e7dbc928d3325b1e7f687086
* AFTER: https://gist.github.com/wesm/ad07cec1613b6327926dfe1d95e7f4f0

IF YOU WANT TO BENCHMARK YOURSELF, PLEASE USE THIS BRANCH https://github.com/wesm/arrow/tree/ARROW-9075-comparison. It contains the RandomArrayGenerator::Boolean change and some other changes to the benchmarks without which the results will be non-comparable