@wesm 2020-06-16
The RTools 4.0 build is spurious. This is ready for review