@wesm 2020-06-16
I implemented some other optimizations, especially for the case where neither values nor filter contain nulls. I'm working on updated benchmarks